Refactor: Move xc_get_user function to apps.plugins.authentication for better organization and backward compatibility
